+++ R/crunchtabs-package.R
@@ -1,0 +1,11 @@
+#' @importFrom kableExtra column_spec kable_styling
+#' @importFrom magrittr `%>%`
+#' @importFrom rlang .data
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL