+++ tests/testthat/test-codeBookItemTxtDescription.R
@@ -7,13 +7,8 @@ with_api_fixture("fixtures-1-2-5", {
     res = codeBookItemTxtDescription(ds$q1)
 
     expect_equal(
-      attributes(res)$kable_meta$contents,
-      c("What is your favorite pet\\?", "")
-    )
-
-    expect_equal(
-      attributes(res)$kable_meta$align_vector_origin,
-      c("l")
+      res,
+      "\\vskip 0.10in\nWhat is your favorite pet?\n\\addcontentsline{lot}{table}{\\parbox{1.8in}{\\ttfamily{q1}} Pet}\n\\vskip 0.10in"
     )
   })
 
@@ -21,59 +16,47 @@ with_api_fixture("fixtures-1-2-5", {
     res = codeBookItemTxtDescription(ds$petloc)
 
     expect_equal(
-      attributes(res)$kable_meta$contents,
-      c("Name the kinds of pets you have at these locations.", "")
+     res,
+     "\\vskip 0.10in\nName the kinds of pets you have at these locations.\n\\addcontentsline{lot}{table}{\\parbox{1.8in}{\\ttfamily{petloc}} Pets by location}\n\\vskip 0.10in"
     )
 
-    expect_equal(
-      attributes(res)$kable_meta$align_vector_origin,
-      c("l")
-    )
   })
 
   test_that("codeBookItemTxtDescription MultipleResponseVariable", {
     res = codeBookItemTxtDescription(ds$allpets)
 
     expect_equal(
-      attributes(res)$kable_meta$contents,
-      c(
-        "Do you have any of these animals as pets\\? Please select all that apply.",
-        ""
-        )
+      res,
+      "\\vskip 0.10in\nDo you have any of these animals as pets? Please select all that apply.\n\\addcontentsline{lot}{table}{\\parbox{1.8in}{\\ttfamily{allpets}} All pets owned}\n\\vskip 0.10in"
     )
 
-    expect_equal(
-      attributes(res)$kable_meta$align_vector_origin,
-      c("l")
-    )
   })
 
   test_that("codeBookItemTxtDescription NumericVariable", {
     res = codeBookItemTxtDescription(ds$ndogs)
 
     expect_equal(
-      attributes(res)$kable_meta$contents,
-      c("No question text", "")
+      res,
+      "\\vskip 0.10in\n\n\\addcontentsline{lot}{table}{\\parbox{1.8in}{\\ttfamily{ndogs}} Number of dogs}\n\\vskip 0.10in"
     )
 
-    expect_equal(
-      attributes(res)$kable_meta$align_vector_origin,
-      c("l")
-    )
   })
 
   test_that("codeBookItemTxtDescription DatetimeVariable", {
     res = codeBookItemTxtDescription(ds$wave)
 
     expect_equal(
-      attributes(res)$kable_meta$contents,
-      c("No question text", "")
+      res,
+      "\\vskip 0.10in\n\n\\addcontentsline{lot}{table}{\\parbox{1.8in}{\\ttfamily{wave}} Wave}\n\\vskip 0.10in"
     )
+  })
 
+  test_that("codebookItemTxtDescription notes", {
+    res = with_mock(codeBookItemTxtDescription(ds$q1), "crunch::notes" = function(x) "This is a note!")
     expect_equal(
-      attributes(res)$kable_meta$align_vector_origin,
-      c("l")
-    )
+      res,
+      "\\vskip 0.10in\nWhat is your favorite pet?\n\\addcontentsline{lot}{table}{\\parbox{1.8in}{\\ttfamily{q1}} Pet}\n\\vskip 0.10in\n\\emph{This is a note!}\n\\vskip 0.10in")
+
   })
 
 })